Search events

+++ src/app/dashboard/widgets/workspace/components/events-correct-card/events-correct-card.component.ts
@@ -13,6 +13,7 @@ import { EventsWorkspaceService } from '../../../../services/widgets/events-work
 })
 export class EventsCorrectCardComponent implements OnInit {
     @Input() public event: IRetrievalEventDto = null;
+    @Input() public isClickable: boolean = true;
 
     public readonly statusesColors: { [id in EventsWidgetNotificationStatus]: string } = {
         new: 'standart',
@@ -20,19 +21,35 @@ export class EventsCorrectCardComponent implements OnInit {
         closed: 'danger',
     };
 
-    constructor(public ewService: EventsWorkspaceService) {}
+    constructor(public ewService: EventsWorkspaceService) { }
 
-    ngOnInit(): void {}
+    ngOnInit(): void { }
 
     public onClickDelete(): void {
+        if (!this.isClickable) {
+            return;
+        }
         this.ewService.deleteRetrievalEvent(this.event);
     }
 
+    public onClickUnlink(): void {
+        if (!this.isClickable) {
+            return;
+        }
+        this.ewService.deleteRetrievalLink(this.event.innerNotificationId);
+    }
+
     public onClickEdit(): void {
+        if (!this.isClickable) {
+            return;
+        }
         this.ewService.editEvent(this.event.innerNotificationId);
     }
 
     public onClickCard(): void {
+        if (!this.isClickable) {
+            return;
+        }
         this.ewService.editEvent(this.event.innerNotificationId);
     }
